cange full filter action
add a Filter Activity and FilterViewModel
change MainViewModel
change all Fragments code
