+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,6 @@
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
     <uses-permission android:name="com.android.vending.BILLING" />
 
-
     <application
         android:name=".AnlikDepremlerApp"
         android:allowBackup="true"
@@ -26,11 +25,13 @@
         <meta-data
             android:name="com.google.android.gms.ads.APPLICATION_ID"
             android:value="${API_KEY_ADMOB}" />
-
         <meta-data
             android:name="com.google.android.maps.v2.API_KEY"
             android:value="${API_KEY_LOCATION}" />
 
+        <activity
+            android:name=".presentation.ui.filter.FilterActivity"
+            android:exported="false" />
         <activity
             android:name=".presentation.ui.language.LanguageActivity"
             android:exported="false" />
