+++ app/src/main/java/com/ferdidrgn/anlikdepremler/presentation/ui/mapsEarthquake/NowEarthquakeAdapter.kt
