+++ app/src/main/res/layout/item_now_earthquake_maps.xml
@@ -46,7 +46,7 @@
                     android:layout_height="85dp">
 
                     <com.google.android.material.card.MaterialCardView
-                        style="@style/CardBackground_primary3"
+                        style="@style/CardBackgroundTint_primary3"
                         android:layout_width="75dp"
                         android:layout_height="75dp"
                         android:layout_centerInParent="true"
