+++ app/src/main/res/values/themes.xml
@@ -78,7 +78,7 @@
         <item name="android:background">@color/primary</item>
     </style>
 
-    <style name="CardBackground_primary3">
+    <style name="CardBackgroundTint_primary3">
         <item name="android:backgroundTint">@color/primary3</item>
     </style>
